# openapi.yaml
openapi: 3.1.0
info:
  title: Aurea Orchestrator API
  description: Automated Unified Reasoning & Execution Agents Orchestrator
  version: 1.0.0
paths:
  /:
    get:
      summary: Root
      description: Root endpoint
      operationId: root__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /request:
    post:
      summary: Create Request
      description: 'Submit a new orchestration request


        This endpoint allows you to submit a task for orchestration. The request will
        be

        queued and await approval before processing.'
      operationId: create_request_request_post
      parameters:
      - name: X-API-Key
        in: header
        required: true
        schema:
          type: string
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrchestrationRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrchestrationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /status/{request_id}:
    get:
      summary: Get Status
      description: 'Get the status of an orchestration request


        This endpoint returns the current status and details of a previously submitted
        request.'
      operationId: get_status_status__request_id__get
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          title: Request Id
      - name: X-API-Key
        in: header
        required: true
        schema:
          type: string
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /approve:
    post:
      summary: Approve Request
      description: 'Approve or reject an orchestration request


        This endpoint allows you to approve or reject a pending orchestration request.

        Approved requests will move to processing status.'
      operationId: approve_request_approve_post
      parameters:
      - name: X-API-Key
        in: header
        required: true
        schema:
          type: string
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    ApprovalRequest:
      properties:
        request_id:
          type: string
          title: Request Id
          description: Unique identifier for the request to approve
        approved:
          type: boolean
          title: Approved
          description: Whether to approve or reject the request
        comment:
          anyOf:
          - type: string
          - type: 'null'
          title: Comment
          description: Optional comment for the approval decision
      type: object
      required:
      - request_id
      - approved
      title: ApprovalRequest
    ApprovalResponse:
      properties:
        request_id:
          type: string
          title: Request Id
          description: Unique identifier for the request
        status:
          $ref: '#/components/schemas/RequestStatus'
          description: Updated status of the request
        message:
          type: string
          title: Message
          description: Response message
      type: object
      required:
      - request_id
      - status
      - message
      title: ApprovalResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    OrchestrationRequest:
      properties:
        task_description:
          type: string
          title: Task Description
          description: Description of the task to orchestrate
        parameters:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Parameters
          description: Task parameters
        priority:
          anyOf:
          - type: integer
          - type: 'null'
          title: Priority
          description: Task priority (1-10)
          default: 1
      type: object
      required:
      - task_description
      title: OrchestrationRequest
    OrchestrationResponse:
      properties:
        request_id:
          type: string
          title: Request Id
          description: Unique identifier for the request
        status:
          $ref: '#/components/schemas/RequestStatus'
          description: Current status of the request
        message:
          type: string
          title: Message
          description: Response message
        created_at:
          type: string
          format: date-time
          title: Created At
          description: Request creation timestamp
      type: object
      required:
      - request_id
      - status
      - message
      - created_at
      title: OrchestrationResponse
    RequestStatus:
      type: string
      enum:
      - pending
      - approved
      - rejected
      - processing
      - completed
      - failed
      title: RequestStatus
    StatusResponse:
      properties:
        request_id:
          type: string
          title: Request Id
          description: Unique identifier for the request
        status:
          $ref: '#/components/schemas/RequestStatus'
          description: Current status of the request
        task_description:
          type: string
          title: Task Description
          description: Description of the task
        created_at:
          type: string
          format: date-time
          title: Created At
          description: Request creation timestamp
        updated_at:
          type: string
          format: date-time
          title: Updated At
          description: Last update timestamp
        result:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Result
          description: Task result if completed
      type: object
      required:
      - request_id
      - status
      - task_description
      - created_at
      - updated_at
      title: StatusResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError